requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
requests
Werkzeug
APScheduler
python-dotenv
Flask-Limiter
redis
gunicorn
